chore: Move common CkShape types out of the utils and into a new file
This makes dependencies cleaner and puts the data in its expected place
