+++ Source/CkShapes/Public/CkShapes/CkShapes_Common.cpp
@@ -1,0 +1,5 @@
+﻿#include "CkShapes_Common.h"
+
+// --------------------------------------------------------------------------------------------------------------------
+
+CK_INTENTIONALLY_EMPTY()
